anim: make Player work with an interface
This is necessary to make the wrappers work correctly, a detail I'd
missed. Unfortunately the interface value will likely result in more
boxing, which is unfortunate if you want to create a lot of these.

I may want to return to the previous version and just require specifying
all the wrappers in the Player type, though that's unfortunate for
tile.Sprite's Player, as it significantly limits its utility and forces
breaking the abstraction.

The good news is that this probably doesn't matter much in practice
because each animation can be pre-allocated in most cases.